requirements.txt
numpy==1.23.1
pandas==1.4.3
psycopg2==2.9.3
SQLAlchemy==1.4.39
requests==2.28.1
beautifulsoup4==4.11.1
python-dateutil==2.8.2   
python-decouple==3.6 